fix(ble/bluedroid): Align config, controller indent and init error paths
- bt_target: remove/align obsolete macros with Kconfig
- device/controller: fix start_up() Secure Connections indent,
  get_ble_resolving_list_max_size return type
- controller.h: align type/interface declarations with implementation
- bte_init: remove unused/redundant code
- bte_main: return -1 on osi_init failure, null check in bte_main_hci_send

(cherry picked from commit b83647f5ea1034b943fd542622dc80363130418a)

+++ components/bt/host/bluedroid/device/controller.c
@@ -546,7 +546,7 @@ static uint8_t get_ble_resolving_list_max_size(void)
     return controller_param.ble_resolving_list_max_size;
 }
 
-static void set_ble_resolving_list_max_size(int resolving_list_max_size)
+static void set_ble_resolving_list_max_size(uint8_t resolving_list_max_size)
 {
     assert(controller_param.readable);
     assert(controller_param.ble_supported);
